% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt2gene.R
\name{apt2gene}
\alias{apt2gene}
\title{Convert SomaScan Analytes to Genes}
\usage{
apt2gene(
  x,
  col_meta_df = NULL,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  collapse = TRUE,
  verbose = interactive()
)
}
\arguments{
\item{x}{Character. A vector of SomaScan analyte identifiers in \code{AptName}
(seq.1234.56) format. Blank ("") values will be coerced to
NA. For more information about SomaScan identifiers and their formats,
please see \link[SomaDataIO:SeqId]{SomaDataIO::SeqId}.}

\item{col_meta_df}{Data frame of column metadata, likely created by
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. This object should contain SomaScan menu
information that can be used to map between \code{AptNames} and Entrez gene
identifiers. At minimum, it must contain columns named "AptName"
(containing SomaScan identifiers in \code{AptName} format) and either
"EntrezGeneSymbol" or "EntrezGeneID".

If not specified, column metadata will be retrieved from \code{example_data_11k}
by default.}

\item{id_type}{Character. Type of gene identifier to be used in the output.
Options are "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol".}

\item{collapse}{Logical. Should values be collapsed to maintain the length of
the original vector? Default is TRUE. Most useful when the results will be
used as a data frame column. When FALSE, all genes associated with a given
\code{AptName} will be returned as individual elements in the output vector. See
examples.}

\item{verbose}{Logical. Should function messages be printed to the console?}
}
\value{
A character vector of genes associated with the values provided to
\code{x}. \code{NA} is returned when a match is not found.
}
\description{
Transform a SomaScan analyte into the gene associated with the analyte's
target.
}
\examples{
anno <- SomaDataIO::getAnalyteInfo(SomaDataIO::example_data)
apts <- withr::with_seed(123,
                         sample(SomaDataIO::getAnalytes(SomaDataIO::example_data), 10))
apts <- c(apts, "seq.10367.62") # Adding IL12
apt2gene(x = apts, col_meta_df = anno)

# Adding results to a data frame
df <- withr::with_seed(321, data.frame(AptName = apts,
                                       Value = rnorm(11)))
df$Gene <- apt2gene(x = df$AptName, col_meta_df = anno)
df

# Using `collapse = FALSE` to separate all returned values
apt2gene(x = df$AptName, col_meta_df = anno, collapse = FALSE)
}
\seealso{
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}
}
\author{
Amanda Hiser, Alex Poole
}
