% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aptPath2genePath.R
\name{aptPath2genePath}
\alias{aptPath2genePath}
\title{Convert SomaScan ID-Based Feature Set into Gene Set}
\usage{
aptPath2genePath(
  path_list,
  col_meta_df,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  verbose = interactive()
)
}
\arguments{
\item{path_list}{Named list containing character vectors of SomaScan
\code{AptNames}.}

\item{col_meta_df}{Data frame of column metadata, likely created by
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. This object should contain SomaScan menu
information that can be used to map between \code{AptNames} and Entrez gene
identifiers. At minimum, it must contain columns named "AptName"
(containing SomaScan identifiers in \code{AptName} format) and either
"EntrezGeneSymbol" or "EntrezGeneID".}

\item{id_type}{Character. Desired gene identifier to be used in the output.
Options include "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol".}

\item{verbose}{Logical. Should progress messages be printed to the console?}
}
\value{
The input \code{pathway} vector, modified to use gene identifiers
instead of \code{AptNames}. \code{AptNames} that do not have an associated gene in
the SomaScan menu (determined using the column metadata in the \code{soma_adat}
object provided to \verb{adat =}) will be dropped.
}
\description{
Utility function to convert a set of SomaScan identifiers (i.e. \code{AptNames})
into genes, while also filtering to avoid potential gene
duplication produced by multiple SomaScan analytes mapping to the same gene.

Mapping between genes and SomaScan analytes is not always 1:1, and it's
possible for multiple analytes to map to the same gene. For example,
\code{13699-6}, \code{4330-4}, and \code{8468-19} all map to prostate-specific antigen
(PSA). If all of these analytes were included in a feature set, PSA would be
represented 3 times in that set. However, when performing over-representation
analysis, pathway analysis, or GSEA, it's crucial to avoid instances where a
single feature is represented multiple times in the same feature set. This
can lead to artificial over-representation when the duplicated feature is
counted multiple times.

After converting SomaScan identifiers to their associated gene, this
function will ensure that each gene is represented only \emph{once} in the gene
set or pathway. However, because 1:many mapping may be present between
identifiers, the pathway returned by \code{aptPath2genePath()} may be longer than
the input aptamer list.
Because of this feature, it is recommended to use this function instead of
\code{apt2gene()} when converting entire gene sets or pathways. See the examples
for more details.
}
\examples{
# Gene symbols are returned by default
aptPath2genePath(path_list   = list(Example = ex_apt_pathway),
                 col_meta_df = SomaDataIO::getAnalyteInfo(example_data_11k))

# When a SomaScan assay analyte is associated with more than one gene,
# all genes will be returned
aptPath2genePath(path_list = list(Single_Analyte = "seq.16927.9"),
                 col_meta_df = SomaDataIO::getAnalyteInfo(example_data_11k))
}
\author{
Amanda Hiser
}
