% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseAptData.R
\name{collapseAptData}
\alias{collapseAptData}
\title{Transform Aptamer-Centric Data to Gene-Centric Data}
\usage{
collapseAptData(
  df,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  sep = "|",
  apt_order = NULL
)
}
\arguments{
\item{df}{Data frame to be converted, most often the output of
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. Must contain, at minimum, an "AptName"
column, and the column specified by \code{id_type}.}

\item{id_type}{Character. Type of gene identifier to be used in the output.
Options are "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol".}

\item{sep}{Character used to separate collapsed SomaScan identifiers.
Default is "|". If NULL, values will \emph{not} be collapsed, and a character
vector will be returned in each column field when >1 SomaScan identifier
is associated with a given gene. See examples.}

\item{apt_order}{Character. The name of a column in \code{df} containing
\strong{numeric} ordering information for collapsed SomaScan identifiers.
By default, no ordering is used. Should only be specified when multiple
aptamers are associated with a given gene, and a particular order is
desired when the values are pasted together into a single, delimited
string.}
}
\value{
A data frame with one row per unique gene.
}
\description{
Utility function to convert a \code{SeqId} or \code{AptName}-centric data frame, with
one row per analyte, to a gene-centric data frame, with one row per gene.
In cases where more than one analyte is associated with a given gene, the
analytes will be collapsed into one row, with a delimiter to separate unique
values.
}
\examples{
df <- SomaDataIO::getAnalyteInfo(example_data_11k)
res <- collapseAptData(df = df)

# Pulling a couple example genes
res[res$EntrezGeneSymbol \%in\% c("NOTCH1", "APOE", "CAMP"), ]

# Specifying a different delimiter
res_delim <- collapseAptData(df, sep = ", ")
res_delim[res_delim$EntrezGeneSymbol \%in\% c("NOTCH1", "APOE", "CAMP"), ]

# Using a ranking metric to reorder the analytes
df$Analyte_Rank <- sample(1:nrow(df), nrow(df), replace = FALSE)
res_ord <- collapseAptData(df, apt_order = "Analyte_Rank")
res_ord[res_delim$EntrezGeneSymbol \%in\% c("NOTCH1", "APOE", "CAMP"), ]

#' Setting `sep = NULL` to create a list-column of SomaScan IDs
res_list <- collapseAptData(df, apt_order = "Analyte_Rank", sep = NULL)
res_list$AptName["APOE"]
}
\author{
Amanda Hiser, Alex Poole
}
