% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2list.R
\name{df2list}
\alias{df2list}
\title{Convert a Data Frame to a Named List of Vectors}
\usage{
df2list(df, name_col, value_col, clean_names = FALSE)
}
\arguments{
\item{df}{Data frame to be converted.}

\item{name_col}{Character. Name of the column in \code{df} to be used for list names.}

\item{value_col}{Character. Name of the column in \code{df} to be used as list
elements (vectors).}

\item{clean_names}{Logical. Should list names be modified to remove
whitespaces and non-aphanumeric strings (excluding underscores)? See
\code{\link[SomaDataIO:cleanNames]{SomaDataIO::cleanNames()}} for more details.}
}
\value{
A list of vectors.
}
\description{
Convert a Data Frame to a Named List of Vectors
}
\examples{
df <- SomaDataIO::rn2col(SomaDataIO::example_data, "Rowname")
df2list(df, name_col = "SampleType", value_col = "Rowname")

# Use clean_names argument to make R-compatible list names
df <- SomaDataIO::getAnalyteInfo(SomaDataIO::example_data)
head(df$TargetFullName)
df2list(df, name_col = "TargetFullName", value_col = "AptName",
        clean_names = TRUE) |>
head()
}
\author{
Amanda Hiser
}
