% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_apt_pathway}
\alias{ex_apt_pathway}
\title{Example Aptamer-Based Pathway}
\source{
\url{gsea-msigdb.org/gsea/msigdb/cards/HALLMARK_INFLAMMATORY_RESPONSE}
}
\description{
Example biological pathway comprised of SomaScan analyte identifiers
(\code{AptNames}). Pathway is a subset of "HALLMARK_INFLAMMATORY_RESPONSE"
(M5932) from MSigDB, converted to \code{AptNames} via \code{gene2apt()}. Only genes
mapping to an analyte in the SomaScan menu were retained.
}
\examples{
head(ex_apt_pathway, n = 20)

length(ex_apt_pathway)
}
\keyword{datasets}
