% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_gene_pathway}
\alias{ex_gene_pathway}
\title{Example Gene-Based Pathway}
\source{
\url{gsea-msigdb.org/gsea/msigdb/cards/HALLMARK_INFLAMMATORY_RESPONSE}
}
\description{
Example biological pathway comprised of gene identifiers (Entrez Gene IDs).
Pathway is a subset of "HALLMARK_INFLAMMATORY_RESPONSE" (M5932) from MSigDB.
}
\examples{
head(ex_gene_pathway, n = 20)

length(ex_gene_pathway)
}
\keyword{datasets}
