% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_11k}
\alias{example_data_11k}
\title{Example 11K Data Set}
\source{
\url{https://github.com/SomaLogic/SomaLogic-Data/blob/main/example_data_v5.0_plasma.adat}
}
\description{
The \code{example_data_11k} object is intended to provide existing and prospective
SomaLogic customers with example data to enable analysis preparation prior
to receipt of SomaScan data, and also for those generally curious about the
SomaScan data deliverable. It is \strong{not} intended to be used as a control
group for studies or provide any metrics for SomaScan data in general.
}
\examples{
# S3 print method
example_data_11k

# Print header info
print(example_data_11k, show_header = TRUE)

# View object class
class(example_data_11k)

# Retrieve annotations
SomaDataIO::getAnalyteInfo(example_data_11k)
}
\keyword{datasets}
