% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2apt.R
\name{gene2apt}
\alias{gene2apt}
\title{Convert Gene Identifiers to SomaScan Analytes (v2)}
\usage{
gene2apt(
  x,
  col_meta_df = NULL,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  collapse = TRUE,
  sep = "|"
)
}
\arguments{
\item{x}{Character. A vector of gene identifiers.}

\item{col_meta_df}{Data frame of column metadata, likely created by
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. This object should contain SomaScan menu
information that can be used to map between \code{AptNames} and Entrez gene
identifiers. At minimum, it must contain columns named "AptName"
(containing SomaScan identifiers in \code{AptName} format) and either
"EntrezGeneSymbol" or "EntrezGeneID".

If not specified, column metadata will be retrieved from \code{example_data_11k}
by default.}

\item{id_type}{Character. Type of gene identifier provided in \code{x}.
Options are "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol".}

\item{collapse}{Logical. Should values be collapsed to maintain the length of
the original vector? Default is TRUE. Most useful when the results will be
used as a data frame column. When FALSE, all genes associated with a given
\code{AptName} will be returned as individual elements in the output vector. See
examples.}

\item{sep}{Character used to separate collapsed identifiers. Default is "|".}
}
\value{
A character vector of SomaScan identifiers (in \code{AptName} format).
}
\description{
Takes a set of gene identifiers and returns associated SomaScan
analytes, given a \code{col_meta} object returned from
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}.
}
\examples{
anno <- SomaDataIO::getAnalyteInfo(example_data_11k)
genes <- withr::with_seed(123, sample(anno$EntrezGeneSymbol, 5))
gene2apt(x = genes, col_meta_df = anno)

# By default, one string for each gene will be returned.
# Genes targeted by >1 analytes will have analytes collapsed via `sep`
gene2apt(x = "IL12B", col_meta_df = anno)

# Use `collapse = FALSE` to return all values individually
gene2apt(x = "IL12B", col_meta_df = anno, collapse = FALSE)
}
\author{
Amanda Hiser, Alex Poole
}
