% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePath2aptPath.R
\name{genePath2aptPath}
\alias{genePath2aptPath}
\title{Convert Gene Set into SomaScan ID-Based Feature Set}
\usage{
genePath2aptPath(
  path_list,
  col_meta_df,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  verbose = interactive()
)
}
\arguments{
\item{path_list}{Named list of character vectors containing genes.}

\item{col_meta_df}{Data frame of column metadata, likely created by
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. This object should contain SomaScan menu
information that can be used to map between \code{AptNames} and Entrez gene
identifiers. At minimum, it must contain columns named "AptName"
(containing SomaScan identifiers in \code{AptName} format) and either
"EntrezGeneSymbol" or "EntrezGeneID".}

\item{id_type}{Character. Type of gene identifier used in \code{path_list}.
Options are "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol".}

\item{verbose}{Logical. Show status messages and (if applicable) a progress
bar?}
}
\value{
List of named character vectors (names = genes, values = AptNames).
When multiple analytes map to the same gene, all are retained as separate
elements with the same name.
}
\description{
Converts gene pathways to SomaScan identifiers. All analytes that map to
a given gene are retained; if multiple analytes map to the same gene, all
will be present in the output.
}
\examples{
genePath2aptPath(path_list = list(Example = ex_gene_pathway), 
                col_meta_df = SomaDataIO::getAnalyteInfo(example_data_11k))
}
\author{
Amanda Hiser
}
