% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt_gene2apt.R
\name{gmt_gene2apt}
\alias{gmt_gene2apt}
\title{Convert Gene-Based GMT File to AptNames}
\usage{
gmt_gene2apt(
  gmt_gene_file,
  col_meta_df,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  write = FALSE,
  gmt_apt_file = NULL,
  verbose = interactive()
)
}
\arguments{
\item{gmt_gene_file}{Path to the \code{.gmt} file to be converted.}

\item{col_meta_df}{Data frame of column metadata, likely created by
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. This object should contain SomaScan menu
information that can be used to map between \code{AptNames} and Entrez gene
identifiers. At minimum, it must contain columns named "AptName"
(containing SomaScan identifiers in \code{AptName} format) and either
"EntrezGeneSymbol" or "EntrezGeneID".}

\item{id_type}{Character. Type of gene identifier used in \code{path_list}.
Options are "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol".}

\item{write}{Logical. Should the converted data set be written out to
a new GMT file? Default is FALSE.}

\item{gmt_apt_file}{The desired file name for the converted file. Required
if \code{write = TRUE}. The file will be written to the current directory,
unless a path to the final file is provided.}

\item{verbose}{Logical. Show status messages and (if applicable) a progress
bar?}
}
\value{
A named list of gene sets/pathways, converted to SomaScan
\code{AptNames}.
}
\description{
Reads in a standard GMT file (using gene identifiers) and converts genes
to SomaScan \code{AptNames}.
}
\examples{
\dontrun{
example_gmt <- system.file("extdata", 
                           "msigdb_h_androgen_response.gmt", 
                            package = "SomaEnrich")
gmt_gene2apt(example_gmt, 
             col_meta_df = SomaDataIO::getAnalyteInfo(example_data_11k), 
             gmt_apt_file = "msigdb_h_androgen_response_apt.gmt")
}
}
