% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go2apt.R
\name{go2apt}
\alias{go2apt}
\title{Retrieve SomaScan Analytes in a Given GO Term}
\usage{
go2apt(x, col_meta_df = NULL, verbose = interactive())
}
\arguments{
\item{x}{Character. A single GO term ID. Must contain "GO:" prefix.}

\item{col_meta_df}{Data frame of column metadata, likely created by
\code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}}. This object should contain SomaScan menu
information that can be used to map between \code{AptNames} and Entrez gene
identifiers. At minimum, it must contain columns named "AptName"
(containing SomaScan identifiers in \code{AptName} format) and either
"EntrezGeneSymbol" or "EntrezGeneID".

If not specified, column metadata will be retrieved from \code{example_data_11k}
by default.}

\item{verbose}{Logical. Should function messages be printed to the console?}
}
\value{
Character vector of SomaScan analyte identifiers in \code{AptName}
format. All analytes associated with each gene are returned. For more
information about SomaScan identifiers and their formats, please see
\link[SomaDataIO:SeqId]{SomaDataIO::SeqId}.
}
\description{
Retrieve SomaScan Analytes in a Given GO Term
}
\details{
A wrapper around \code{\link[=path2apt]{path2apt()}} specifically for GO terms.
}
\examples{
go2apt("GO:0019319") # Returns vector of AptNames

# Apply over a vector of GO terms
sapply(c("GO:2001256", "GO:0000462", "GO:2001259"), go2apt, USE.NAMES = TRUE)
}
