% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaORA.R
\name{somaORA}
\alias{somaORA}
\title{Perform Overrepresentation Analysis for SomaScan}
\usage{
somaORA(
  features,
  universe,
  resource = c("bp", "mf", "h", "c1", "c3", "c4", "c6", "c7", "c8"),
  split_ids = TRUE,
  unique_features = TRUE,
  id_type = c("EntrezGeneSymbol", "EntrezGeneID"),
  use_aptnames = FALSE,
  col_meta_df = NULL,
  cust_paths = NULL,
  min_feats = 5L,
  max_feats = 500L,
  verbose = interactive()
)
}
\arguments{
\item{features}{Character. A vector containing features of interest, often
differentially expressed genes or SomaScan analytes. Vector elements must
be either Entrez gene symbols/identifiers (unless custom gene sets are
used, see \code{cust_paths}) or SomaScan identifiers in R-compatible
\code{AptName} format (e.g. \code{seq.1234.56}). If using \code{AptNames}, you must set
\code{use_aptnames = TRUE}.}

\item{universe}{Character. A vector of unique gene symbols or \code{AptNames}.
This vector represents the overall "universe" from which the elements in
\code{features} were identified. The \code{universe} should encompass \emph{all}
elements in the differential expression analysis from which the set of
over-represented targets (\code{features}) was chosen. See \code{Details} for more
information.
If using \code{AptNames}, you must set \code{use_aptnames = TRUE}.}

\item{resource}{Character. String specifying the gene set database resource
to use. Choose one of:
\describe{
\item{"bp"}{GO Biological Process}
\item{"mf"}{GO Molecular Function}
\item{"h"}{MSigDB Hallmark}
\item{"c1"}{MSigDB Positional}
\item{"c2"}{MSigDB Curated (note: does not contain KEGG or BioCarta)}
\item{"c3"}{MSigDB Regulatory Target}
\item{"c4"}{MSigDB Computational (note: does not contain CM submodule)}
\item{"c6"}{MSigDB Oncogenic Signature}
\item{"c7"}{MSigDB Immunologic Signature}
\item{"c8"}{MSigDB Cell Type Signature}
}
The default is \code{bp}. For more advanced filtering, or to use gene sets not
provided here, see the \code{cust_paths} argument. Note that \code{resource} is
ignored when a data set is supplied to \code{cust_paths}.}

\item{split_ids}{Logical. If the identifiers used in the \code{features} and
\code{universe} vectors contain heterodimers (elements separated by a
delimiter, like \code{"GENE1|GENE2"}), should the heterodimers be split into
individual components, like \code{c("GENE1", "GENE2")}? If TRUE (the default),
this may introduce duplicate identifiers into \code{features}/\code{universe}.
Applies to \emph{both} \code{features} and \code{universe}. See \code{Details} and examples.}

\item{unique_features}{Logical. If present, should duplicates in
\code{features} be removed? Default is FALSE. If TRUE, the resultant
vector will only contain unique elements. The duplicated features are
often the result of setting \code{split_ids = TRUE} when heterodimers
are present in the \code{features} vector. See examples.}

\item{id_type}{Character. Type of gene identifier used in \code{features}.
Options are "EntrezGeneSymbol" or "EntrezGeneID". Default is
"EntrezGeneSymbol". Ignored if \code{use_aptnames = TRUE}.}

\item{use_aptnames}{Logical. Should analysis be performed in aptamer-space,
i.e. using SomaScan \code{AptNames} instead of genes? Default is FALSE. If TRUE,
the input ranking data must contain \code{AptNames} instead of gene
identifiers. If using \code{cust_paths} when \code{use_aptnames = TRUE}, custom
gene set members must also be \code{AptNames}.}

\item{col_meta_df}{Optional. Data frame of column metadata, likely created
by calling \code{\link[SomaDataIO:getAnalyteInfo]{SomaDataIO::getAnalyteInfo()}} on the \code{soma_adat} used for
differential expression analysis. Required when
\code{use_aptnames = TRUE}. This object will be used to map between SomaScan
\code{AptNames} and gene symbols.}

\item{cust_paths}{Optional. A named list containing character vectors of
gene sets or functional groups of interest. If provided, \code{cust_paths} will
override \code{resource}. The name of each list element should be the name of
the pathway/gene set. The identifier used in each vector should be
the same type as the names of \code{ranks}. This object is still subject to the
minimum/maximum feature cutoffs defined by \code{min_feats} and \code{max_feats}.

This argument is meant to be used when the desired biological network is
not present in the \code{resource} options.}

\item{min_feats}{Numeric. Minimum number of features required for a gene set
to be retained. Default is 5. Sets smaller than this value will be
discarded. If \code{NULL}, no minimum is used. See \code{Details} for more
information.}

\item{max_feats}{Numeric. Maximum number of features required for a gene set
to be retained. Default is 500. See \code{Details} for more information.}

\item{verbose}{Logical. Should progress messages be printed to the console?
By default, messages will only be shown in interactive R sessions. Set
to TRUE to show all messages, and FALSE to silence messages.}
}
\value{
An object of class \code{somaORA} (inheriting from \code{data.frame})
containing the results of ORA, where the results in each row correspond
to a single tested pathway. The data frame contains the following columns:
\item{resource_code}{Abbreviated character string representing the name
of the original gene set resource.}
\item{pathway_id}{Pathway identifier/accession number from the original
gene set or pathway resource.}
\item{pathway}{Name of the pathway.}
\item{pval}{An enrichment p-value from hypergeometric test.}
\item{padj}{The BH-adjusted p-value.}
\item{foldEnrichment}{The degree of enrichment, relative to the background.}
\item{overlap}{The size of the overlap between the pathway and features of interest.}
\item{size}{The size of the pathway.}
\item{overlapFeatures}{A vector containing the overlapping features (see 'overlap' column).}

By default, the data frame is sorted by the BH-adjusted p-value and fold
enrichment.
}
\description{
Implements the algorithm from \code{\link[fgsea:fora]{fgsea::fora()}} to perform
overrepresentation analysis (ORA) via hypergeometric test, tailored
specifically to proteomic SomaScan data.
}
\section{Additional Details}{

Overrepresentation analysis is a statistical method that determines
whether sets of features, often genes, are present in a data set
more than would be expected by chance. As an example, imagine a SomaScan
experiment in which 200 analytes are determined to be differentially
expressed in a group of breast cancer samples. Using ORA, a
researcher can investigate whether any pathways, like cellular
proliferation, are more highly represented in those 200 analytes than
would be expected by chance.
}

\section{Feature Selection}{

The primary input to ORA is a set of biological features of interest.
Selecting these features typically involves defining a specific subset,
often differentially expressed genes, based on previously defined
thresholds or cutoffs. For example, one could use the combined statistical
significance and log fold change (logFC) cutoffs of p < 0.05 and
logFC > 2, respectively, to identify significantly differentially
expressed genes. This list would then be used as input for ORA. Please
note that the most appropriate cutoff(s) to use will likely vary by
experiment; this is only an example.
}

\section{Enrichment Universe}{

The \code{universe} (sometimes referred to as "background") serves as a point
of reference against which the genes or analytes of interest (provided to
\code{features}) can be compared. Setting the correct background universe is
\emph{crucial} for accurate ORA results. The universe must encompass all
features in the analysis results (typically differential expression
analysis). If analyte filtering was performed prior to differential
expression analysis, the filtered analytes should not be used to set the
ORA universe.

If \code{features} contains elements that are not in \code{universe}, those
elements will be dropped from \code{features}. Additionally, \code{universe}
elements that are not found in any gene set will be dropped.
}

\section{Handling Heterodimers}{

When a SOMAmer Reagent targets a protein heterodimer, the genes associated
with that heterodimer are collapsed using a delimiter character
(ex. "CGA|CGB3|CGB7") for concise display in the SomaScan menu. However,
when performing ORA, these collapsed and delimited strings must be split
into individual gene symbols, so each symbol can be matched against
gene-based databases like GO or MSigDB. In SomaEnrich, the \code{resource} or
\code{cust_paths} arguments direct annotation of the features of interest to
known pathways or gene sets. If the elements of \code{features} do not match
elements in \code{resource}/\code{cust_paths}, the non-matching elements in
\code{features}/\code{universe} will be dropped. Consequently, to preserve these
elements, the default behavior is to split delimited strings.
}

\section{Min/Max Set Sizes}{

The statistical power of GSEA/ORA to detect enrichment depends on the size
of the gene set being tested, and GSEA normalization is not accurate
for very small feature sets. The minimum allowable gene set is defined by
the \code{min_feats} argument; any sets smaller or larger than \code{min_feats} or
\code{max_feats} will be discarded \emph{before} performing GSEA/ORA. Please note
that this threshold is applied \emph{after} filtering to remove features from
each gene set that are not also present in the \code{ranks}, i.e. taking the
intersect of the \code{ranks} and gene lists.
The default \code{min_feats} threshold used here differs from other enrichment
analysis tools. The 15-gene minimum suggested by the Broad Institute's
\href{https://www.gsea-msigdb.org/gsea/index.jsp}{GSEA software} was established
in a whole-genome context, where the previously described intersection step
removes few genes. However, in the reduced feature space of SomaScan,
more genes are lost from taking this intersection. In some cases, this can
greatly reduce the size of available gene sets. To account for this loss, a
default minimum threshold of 5 is used. Consider final gene set size
accordingly when reviewing your enrichment analysis results.
}

\examples{
# Genes mapping to differentially expressed analytes
deg <- t_tests[t_tests$fc > 1 & t_tests$fdr < 0.05, ]$EntrezGeneSymbol

# Input universe = all genes in statistical test results
uni <- t_tests$EntrezGeneSymbol

# To properly check for enrichment, values in the input vector must match 
# values in the gene sets/pathway resource. 
# SomaScan contains heterodimers that won't match non SomaScan-derived gene sets
head(deg)
deg[grepl("\\\\|", deg)]

# Splitting these values can create duplicates in the input vector
split_deg <- strsplit(deg, "\\\\|") |> unlist()
split_deg[duplicated(split_deg)]

# The following will run ORA with GO BP gene sets (the default),
# split heterodimers into individual components (the default),
# and remove duplicates created from splitting
res <- somaORA(features = deg,
               universe = uni,
               unique_features = TRUE)

# ORA can be performed using SomaScan AptNames, instead of genes
dea <- head(t_tests$AptName, 50)
apt_uni <- t_tests$AptName

# A progress bar will appear to track AptName conversion
meta <- SomaDataIO::getAnalyteInfo(example_data_11k)
res_apt <- somaORA(features = dea,
                   use_aptnames = TRUE,
                   universe = apt_uni,
                   col_meta_df = meta)
}
\author{
Amanda Hiser
}
