% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_tests}
\alias{t_tests}
\title{t-test Example Data}
\source{
\url{https://somalogic.github.io/SomaDataIO/articles/stat-two-group-comparison.html}
}
\description{
Results of performing differential expression (via t-test) with an example
SomaScan dataset (\code{example_data_11k}). The \code{Sex} variable was used to compare
males (N = 81) against females (N = 81). The results are sorted by
t-statistic and adjusted p-value.
}
\details{
The SomaDataIO "Two-Group Comparison" workflow was used to produce this data
set, with modifications (different input ADAT, and fold change and
log2 fold change values were added to the final table).
}
\examples{
colnames(t_tests)

head(t_tests)

# Formula used to generate results for a given analyte
t_tests$formula[[1]]
}
\keyword{datasets}
